--- api/src/main/java/org/example/age/api/LiteHttpHandler.java
@@ -1,8 +1,0 @@
-package org.example.age.api;
-
-/** Lightweight HTTP handler that can dispatch requests and send a response. */
-@FunctionalInterface
-public interface LiteHttpHandler<S extends Sender> {
-
-    void handleRequest(S sender, Dispatcher dispatcher) throws Exception;
-}
